qcacld-3.0: Not clear active_links for multi link sap SSR
For multi link sap, active_links is used to get how
many links have been started which will be used for SSR
and rmmod. which should not be cleared.

For SSR the call is :
hdd_reset_all_adapters->hdd_stop_adapter->
->hdd_stop_adapter_ext->hdd_adapter_disable_all_links

For rmmod, hdd_deinit_adapter happen after this function,
active_links still be used, so not clear it, call flow is:
	hdd_stop_adapter->hdd_stop_adapter_ext->
		hdd_adapter_disable_all_links
	hdd_deinit_adapter

Change-Id: I9c214fafc17726b6de0224be8deacba334c76267
CRs-Fixed: 3720084