qcacld-3.0: Check against self vdev id for IPA TX pipe
It's observed that for STA vdev, when getting vdev id for connection
with same frequency, vdev id of its own is returned back to HDD IPA
layer. This is caused due to updating connection table in policy_mgr
component for the STA vdev takes place before getting the IPA TX pipe
for it in HDD IPA layer. In other words, providing only the new
frequency for the target connection is not sufficient.

Fix is to provide the self vdev id for the connection with new frequency
and return valid vdev id only when it's not vdev of it own.

CRs-Fixed: 3735788
Change-Id: If98a9a7a5d223b05f55f50daba174ea6fe5ed3dd