qcacld-3.0: Deinit QMI WFDS client prior to WFDS work destroy
As part of idle shutdown QMI WFDS workqueue and work
are destroyed prior to WFDS QMI client deinit. This
could result in a possible race condition where wq
and work are destroyed but work is getting queued
as part of WFDS server up.

Fix is to deinitialize the WFDS QMI client prior to
destroying WFDS wq and work.

CRs-Fixed: 3731687
Change-Id: I599e95fc7914cd2c5881b331bc96319d0365b483