qca-wifi: Add TPE vendor commands to support for WFA
Add vendor commands definition to allow user configuration for
1. The type of power unit to be advertised in the TPE IE either
PSD or EIRP.
2. Single common PSD value or multiple PSD values for all 20MHz
subchannels in the current channel of the TPE-IE.
3. To configure the punctured channel's PSD power as -128.
Example: AP is operating on channel 33 HT80 with a puncture pattern of
0x8 (channel 45 is punctured). If the option "cfg80211tool athx
set_tpe_punc_chan_power 1" is set, the driver will set the PSD power of
channel 45 as -128. If this option is not set, the driver will set the
PSD power of the channel 45 as primary20MHzPSD - 16dbm. This option is
used only for WFA validation.

CRs-Fixed: 3655027
Change-Id: I7c95ab0f72d4634fd5af38e931581a87683df4c4