qcacld-3.0: Update max channel switch time value for SAP
The old max channel switch time only include the
CAC timeout if has. But actually, it is not to be
much more accurate. And we should include two beacon
interval and vdev restart time as well.
One beacon interval time is for CSA count 1 reduced
to count 0, and the other beacon interval is required
when send the first beacon on the new channel.

Change-Id: I647b387de54b8fd708f74b98e253c08fbce2456f
CRs-Fixed: 3737362