qcacld-3.0: Fix populating EHT 320 MHz beam formee capability
In the api populate_dot11f_eht_caps(), beam_formee capability
on 320 MHz EHT capability is populated based on FW capability
irrespective of the current channel bandwidth.

Add check to populate beam_formee capability on 320 MHz EHT
capability if the FW supports this capability and the current
connection is established on channel with 320 MHz bandwidth.

Change-Id: I6d8e933e82e8bcd4be4f2c2aadbe5be9244ca0a1
CRs-Fixed: 3737505