qcacld-3.0: Select valid default frequency for ll_lt_sap
Currently, if XPAN SAP comes up immediately after any SAP
interface, in that case XPAN SAP does not do ACS for the
channels which are already scanned as part of the ACS of
previous SAP, If host driver skips all the channels of
XPAN ACS, then it selects any default channel. This
default channel can be SCC or MCC channel with respect
to existing SAP. Since XPAN does not support SCC or MCC
with any other SAP, or it can be a 2 GHz channel, these
all channels are not valid channels for LL_LT_SAP.

To avoid above issue, add a logic to select a valid
channel for XPAN if already a SAP is present.

Change-Id: I24698456d41b50030584e26b197b8ea3c3a2ea9e
CRs-Fixed: 3746188