qca-wifi: Handle stats delay due to error in fetching
The async stats should return immediately in case of an inappropriate
combination for the wifitelemetry command. But due to the timeout,
we see that it waits for about 5 seconds to present the error code.

Fix this by leveraging the error handler in the send_nlmsg to detect
the errors in the driver.

CRs-Fixed: 3768684
Change-Id: Ia1967d3ea00685870a3f9d22506320d371698896