qcacld-3.0: NAN/NDI peer migration
Currently target allows only one peer on one mac address.
And some vendor uses same interface for NAN and NDI. This lead
to same mac address for NAN PASN peer and NDP peer and it causes
crash in the target.

So, to avoid this, driver handles peer migration in below way:
1. On NDP indication or NDP response, driver will check if same
peer exists on NAN or not.
2. If same peer exists on NAN, it will delete the NAN PASN peer
and create the same NDI peer.
3. ON NDP end, it will delete the NDP peer and recreate the same
peer on NAN.

Change-Id: I3286208187fceda698bcc66d0bda1c37316e713b
CRs-Fixed: 3774948