qcacld-3.0: Add optional argument to treat NOL as enabled
The regulatory API 'reg_get_band_from_cur_chan_list' is used
for copying the channels which are enabled from the input
channel list based on the input band flags. Currently this
API treats NOL as disabled and excludes the NOL channels
from the resultant channel list.

Change I49e855a6c86dcc8dfa731d1e7599db242f786c0e updates this
API to have an additional optional argument to pass a flag to
determine whether to treat NOL channels are enabled or disabled.

With this change, add a fix to update the caller APIs of
reg_get_band_from_cur_chan_list to pass a flag as an additional
argument to determine whether to treat NOL channels as enabled
or disabled.

Change-Id: I567287bf798d8b876b0c6fcef940636913f99e2a
CRs-Fixed: 3781972