qcacld-3.0: Fix disconnect request stuck in lim
The STA disconnect request is stuck in lim message queue because
the CM_DISCONNECT_REQ is pending in gLimDeferredMsgQ queue and queue
is paused.
The reason is STA CSA is in progress and new CSA event is received,
the vdev mgr SM is not supported and failed to handle it, that causes
the gLimDeferredMsgQ queue is not resumed.

Fix by skip the CSA event if CSA is in progress.

Change-Id: Ib0e667f153102601183be82e64c74cd46c1fb22b
CRs-Fixed: 3784801