qcacld-3.0: Refine partner links filter for common AKM
Common AKM with partner links fails if the number of entries
from scan filter doesn't equal the total partner links.
If scan DB has duplicate entry with same BSSID then this is
possible.

Instead of downgrading to SLO in such cases, check if the
number of unique entries in the filter equals number of
partner links or not as criteria for downgrade to SLO.

Also add SSID to scan filter.

Change-Id: I2fd8642ea450b75ed3178f926cc43c4c15c5b8e7
CRs-Fixed: 3806681