qcacld-3.0: Max weight to non-ACS channel
Currently, host driver sorts all the channels for particular
channel width based on weight and BSS count, irrespective of
channel is present in ACS list or not.
Sometimes, it happens that non-ACS channels (channels which are
not present in ACS list) have more preference than ACS channels
for that particular channel width in the sorted channel list.

So, to avoid this, add check and assign weight to maximum value
for non-ACS channel. This will provide less preference for
non-ACS channel over ACS channel in the sorted channel list.

Change-Id: I6bca3f38c68687031bcf5abb09ed427947ba4d9d
CRs-Fixed: 3772913