qcacld-3.0: Update ACS result without unsafe channels
Consider the case when there is only 1 valid 80 MHz channel
available and some of those bonded channels are marked unsafe.
In this case, even though the unsafe weightage is high, it gets
selected since it is the only channel available. However, upon
SAP start since the channel is unsafe, the driver rejects SAP.

To fix this, if one of the bonded channel for a given channel
width is unsafe, host has to mark the entire range as invalid.
If no valid channel is present for a given range, then ACS
logic will fallback to next lower bandwidth.

Change-Id: I826f266c56d56286336c097900fff675d427f2f7
CRs-Fixed: 3787283