qcacld-3.0: Do not handle zero ACS results if LL_LT_SAP is present
Currently host driver handles zero ACS results if LL_LT_SAP is
present and it fills some valid frequency from ACS list if there
is no valid frequency remains in this ACS list after host driver
filters the ACS list based on the current allowed frequencies.

Current issue is that if LL_LT_SAP is present on static sbs device
and along with it if another SAP tries to come up on 5 GHz which
contains only 5 GHz frequencies in ACS request, host driver filters
all these 5 GHz frequencies, as LL_LT_SAP and SAP is not supported
in MCC. Once host driver makes number of ACS frequencies as zero,
it tries to handle the zero ACS frequency case and it fills one
valid frequency and returns that frequency as ACS result.
Now, when SAP start comes up, host driver handles this start ap
and starts SAP on some valid 2.4 GHz frequency as per the current
concurrencies.
After sometime if DCS riggers on the SAP and as part of this
host driver tries to change the channel, to find the new channel
it runs ACS again and tries to find a new channel. Since ACS
master list contains only one valid channel (that is 5 GHz channel),
so it selects that channel and starts SAP on that channel, this
results in MCC with LL_LT_SAP which is not supported by FW.

To address above issue, do not return any channel once ACS list
becomes zero channel list based on LL_LT_SAP concurrencies as
this list becomes zero based on LL_LT_SAP concurrencies.

Change-Id: I8f6d5fccce963f269b1e505ee2e57aa51a9a2ef0
CRs-Fixed: 3784600