qcacld-3.0: Avoid TWT setup on 2.4 GHz when enable_twt_24ghz ini is 0
In the api osif_twt_setup_req(), TWT setup command is processed
successfully when connected to 2.4 GHz band and ini enable_twt_24ghz
is set to '0'.

Add check in the api osif_twt_setup_req() to prevent TWT
setup when connected to 2.4 GHz band and ini enable_twt_24ghz
is set to '0'.

CRs-Fixed: 3784042
Change-Id: I086f38d1f10e384141a3e8cc8b65b1fa52827a74