qcacld-3.0: Remove entry from sampling table from single context
Currently an entry can be removed from the sampling table
via two different contexts (STC periodic work and the STC sampling
timer), which can lead to race conditions between the count of
the flows in sampling table and the sampling entry state.
Such a scenario can result in access to invalid sampling table
entry.

In order to mitigate this issue, limit the removal of the
sampling table entry within single context, which serializes
the entry removal and avoid any race condition.

Change-Id: Icf534239683ae56cd4abf6405137068265fe036a
CRs-Fixed: 3823931