qcacld-3.0: Reconfigure KEEPALIVE interval after link switch
As part of SET_KEEP_ALIVE_INTERVAL command host configures
KEEPALIVE interval for connected ml vdevs and resets to
default value as part of disconnect. In case of disconnect
as part of link repurpose host resets KEEPALIVE interval
and fails to set in connect case.

To address this issue, don't reset KEEPALIVE interval
incase of link repurpose and set it after connection.

Change-Id: Ia62109ceb32740844cfef3ef975ab15492e64b8c
CRs-Fixed: 3805788