qcacld-3.0: Free allocated memory for nan disable request properly
Currently in the function ucfg_disable_nan_discovery(), if all
nan peers are not deleted successfully we return from the function
not freeing the memory allocated to nan_disable_req. Thus causing a
memory leak. So, to avoid this issue, free the allocated memory
properly before returning from the function.

Change-Id: I7c6e9949c52c206b5979013748e7bf4909528306
CRs-Fixed: 3822490