qcacld-3.0: Clear candidate entry flags after moving to sampling table
Currently the candidate entry flags are not cleared after it
is moved to sampling table. Due to this uncleared flags, it is
possible that these candidate entries are again interpreted as
valid entries during the next iteration.

Hence to mitigate this issue, clear the candidate entry flags
after moving it to sampling flow table.
Also remove the assert when handling flow classify result.

Change-Id: I613a382b038db73a5942d439eba02e4fbe2bf339
CRs-Fixed: 3825034