qcacld-3.0: Add duplicate peer detection logic before TDLS setup
Issue #1:
When Legacy sta does TDLS with one link of ML Sta and then
moves to the other link AP due to roaming or other reason, then
the same STA can try TDLS on the other link.
While processing TDLS add sta, peer create is done. If duplicate
peer exists, then WMA returns error and add sta failure status
is given to userspace. But the allocated peer idx and sta_ds
are not cleaned up resulting in TDLS peer delete sent on
wrong vdev during disconnect because the peer delete is done
for the pe_session for all the AID present in the aid bitmap
that is stored in the PE session.
In this failure scenario peer object already exists but sta_ds
is not present.

Add duplicate peer detection in lim while checking for sta_ds
also to avoid cleanup failure & add proper error handling
in lim_process_tdls_add_sta_rsp() for failure case.

Issue #2:
When Link Switch happens, as part of the disconnect sequence
the TDLS peers are not deleted sometimes. This is because, TDLS
module posts delete all peers msg to lim, but the disconnect
defers other msgs and pe_session is deleted. When TDLS msg gets
processed, it checks for pe_session to delete all the peers
based on the peer AID bitmap stored in pe_session.
Since PE session was deleted, the peers are not deleted and AID
was also cleared. So the TDLS peers where not deleted during
subsequent disconnections also.

Delete all TDLS peers directly when disconnect start indication
is received in IF MGR without posting msg, similar to south
bound disconnect & roaming cases.

Change-Id: I403fe0022289d4a5d5b45fdd37477b89ce50ec1e
CRs-Fixed: 3823658