qcacld-3.0: Fix incorrect peer_id in STC for bi-di flows
Currently the peer_id is not getting updated in the
sampling table when the shortlisted flow is a bi-directional
flow. Due to this, the active traffic map indication is not
sent to FW even after the flow has been classified.

Fix this by saving the peer_id in case of bi-directional
shortlisted flow for STC.

Change-Id: I925efb7e3d4e6bec790cb7a9498cb73ddee4446a
CRs-Fixed: 3829760