qcacld-3.0: Reset the selected_to_sample flag after classification
Currently the "selected_to_sample" flag is not being reset after the
flow gets classified. In a scenario where one of the flows (Tx or Rx)
from a bi-directional flow gets retired, and subsequently both the
flow are marked as inactive. Now if the non-retired flow of the above
mentioned bi-directional pair resumes traffic, it is never selected to
be sampled/classified.

In order to fix the above mentioned issue, reset the
selected_to_sample flag in the Tx/Rx flow entries after
the flow has been classified.

Change-Id: Idd0d7d3cda54233307e78ec026a52d182205e7b5
CRs-Fixed: 3830031