qcacld-3.0: Optimize connect req path
Currently, when user send join request to the driver, it post
message to LIM layer. In LIM, join request is processed and peer
is created and it waits for peer confirmation. On peer
confirmation, CDP peer is setup and confirmation are sent to the
connection manager. Connection manager send connect request to LIM
layer which send VDEV start request to the target.
And cdp_peer_setup is also posting message to the schedular thread
to send WMI_PEER_REORDER_QUEUE_SETUP_CMDID. The scheduler is
processing the cdp_peer_setup getting messages before connection
message. Thus, WMI_PEER_REORDER_QUEUE_SETUP_CMDID is sent before
VDEV start. This leads to delay in the connection.

So, to fix this, exempt the connect request ("CM_CONNECT_REQ")
message to the scheduler and calls handle join request API
directly from the connection manager.

Change-Id: If31c3af0c0983822f58869a07f0b7ed462755505
CRs-Fixed: 3769806