qcacld-3.0: Consider NAN disc channels to handle NAN+P2P+ML-STA
When ML-STA is present in 5GL + 5GH, it can be in MLMR if it's
SBS supported target. Below actions are taken when NAN comes up,
1. Disable/disallow eMLSR as part of NAN pre-enable
2. Disable/disallow MLMR between 5GL and 5GH.

Same holds good even when NAN is present and ML-STA comes up later.

Follow the same for disallow_modes(eMLSR and MLMR) when
NAN is present/NAN comes up.
Consider both 2 GHz and 5 GHz NAN discovery frequencies
to evaluate the disallowed modes.

Change-Id: Icf7e23835e89de51ca54c193062b8e8600135592
CRs-Fixed: 3827348